clients/python/requirements.txt
requests>=2.31
websockets>=15.0
websocket-client>=1.6.0
aiohttp>=3.9.0
scipy>=1.11.0
numpy>=1.24.0
matplotlib>=3.7.0
astral>=3.2
python-rtmidi>=1.5.0
pyaudio>=0.2.14
sounddevice>=0.4.6
samplerate>=0.1.0
pyinstaller>=6.17.0
zeroconf>=0.132.0
pyserial>=3.5
opuslib>=3.0.1
zstandard>=0.22.0

# Windows-only: OmniRig support
pywin32>=306; platform_system=="Windows"
